refactor(navigation): merge Link component into typography atoms
